requirements.txt
anyio==3.6.1
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
datastore-entity==0.2.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
firebase-admin==5.2.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
google-api-core==2.7.2
google-api-python-client==2.51.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-datastore==2.7.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.22.3
orjson==3.7.2
packaging==21.3
pandas==1.4.2
proto-plus==1.20.6
protobuf==3.20.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing-extensions==4.2.0
ujson==5.3.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.0
